@@ -1,17 +1,17 @@
 package com.adaptris.core.jcsmp.solace;
 
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 
 import com.adaptris.core.AdaptrisMessage;
@@ -24,7 +24,7 @@ import com.solacesystems.jcsmp.BytesXMLMessage;
 import com.solacesystems.jcsmp.JCSMPException;
 import com.solacesystems.jcsmp.JCSMPFactory;
 import com.solacesystems.jcsmp.JCSMPSession;
-import com.solacesystems.jcsmp.JCSMPStreamingPublishEventHandler;
+import com.solacesystems.jcsmp.JCSMPStreamingPublishCorrelatingEventHandler;
 import com.solacesystems.jcsmp.Queue;
 import com.solacesystems.jcsmp.XMLMessageProducer;
 
@@ -56,7 +56,7 @@ public class SolaceJcsmpQueueProducerTest extends MockBaseTest {
 
   @Mock private BytesXMLMessage mockMessage;
 
-  @Before
+  @BeforeEach
   public void setUp() throws Exception {
     adaptrisMessage = DefaultMessageFactory.getDefaultInstance().newMessage();
 
@@ -77,7 +77,7 @@ public class SolaceJcsmpQueueProducerTest extends MockBaseTest {
     .thenReturn(mockQueue);
     when(mockConnection.retrieveConnection(SolaceJcsmpConnection.class))
     .thenReturn(mockConnection);
-    when(mockSession.getMessageProducer(any(JCSMPStreamingPublishEventHandler.class)))
+    when(mockSession.getMessageProducer(any(JCSMPStreamingPublishCorrelatingEventHandler.class)))
     .thenReturn(mockProducer);
     when(mockTranslator.translate(adaptrisMessage))
     .thenReturn(mockMessage);
@@ -85,7 +85,7 @@ public class SolaceJcsmpQueueProducerTest extends MockBaseTest {
     LifecycleHelper.initAndStart(producer);
   }
 
-  @After
+  @AfterEach
   public void tearDown() throws Exception {
     LifecycleHelper.stopAndClose(producer);
   }
@@ -122,7 +122,7 @@ public class SolaceJcsmpQueueProducerTest extends MockBaseTest {
 
   @Test
   public void testMessageProducerCache() throws Exception {
-    when(mockSession.getMessageProducer(any(JCSMPStreamingPublishEventHandler.class)))
+    when(mockSession.getMessageProducer(any(JCSMPStreamingPublishCorrelatingEventHandler.class)))
     .thenReturn(mockProducer)
     .thenReturn(mockProducer2);
 
