@@ -1,6 +1,7 @@
 package com.adaptris.core.jcsmp.solace;
 
 import javax.validation.constraints.NotBlank;
+
 import com.adaptris.annotation.AdapterComponent;
 import com.adaptris.annotation.ComponentProfile;
 import com.adaptris.annotation.DisplayOrder;
@@ -12,6 +13,7 @@ import com.solacesystems.jcsmp.FlowReceiver;
 import com.solacesystems.jcsmp.JCSMPSession;
 import com.solacesystems.jcsmp.Queue;
 import com.thoughtworks.xstream.annotations.XStreamAlias;
+
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -19,7 +21,8 @@ import lombok.Setter;
 
 /**
  * <p>
- * This implementation of {@link AdaptrisMessageConsumer} will use the Solace Jcsmp Api to consume messages from a Queue on your Solace router.
+ * This implementation of {@link AdaptrisMessageConsumer} will use the Solace Jcsmp Api to consume messages from a Queue on your Solace
+ * router.
  * </p>
  * <p>
  * There are four main components that you will need to configure;
@@ -30,14 +33,15 @@ import lombok.Setter;
  * <li><b>Acknowledge mode: </b> Should either be CLIENT or AUTO.</li>
  * </ul>
  * </p>
+ *
  * @author aaron
  * @config solace-jcsmp-queue-consumer
  */
 @AdapterComponent
-@ComponentProfile(summary="A Solace native JCSMP component that consumes your messages from the Solace VPN.", tag="queue,consumer,solace,jcsmp", since="3.9.3")
+@ComponentProfile(summary = "A Solace native JCSMP component that consumes your messages from the Solace VPN.", tag = "queue,consumer,solace,jcsmp", since = "3.9.3")
 @XStreamAlias("solace-jcsmp-queue-consumer")
 @NoArgsConstructor
-@DisplayOrder(order = {"queue", "acknowledgeMode", "endpointPermissions", "endpointAccessType"})
+@DisplayOrder(order = { "queue", "acknowledgeMode", "endpointPermissions", "endpointAccessType" })
 public class SolaceJcsmpQueueConsumer extends SolaceJcsmpAbstractConsumer {
 
   @Getter(AccessLevel.PACKAGE)
@@ -53,7 +57,6 @@ public class SolaceJcsmpQueueConsumer extends SolaceJcsmpAbstractConsumer {
   @NotBlank
   private String queue;
 
-
   @Override
   public void startReceive() throws Exception {
     getSessionHelper().createSession();
@@ -68,19 +71,20 @@ public class SolaceJcsmpQueueConsumer extends SolaceJcsmpAbstractConsumer {
 
   @Override
   public void stop() {
-    if(getFlowReceiver() != null)
+    if (getFlowReceiver() != null) {
       getFlowReceiver().stop();
+    }
     super.stop();
   }
 
   @Override
   public void close() {
-    if(getFlowReceiver() != null)
+    if (getFlowReceiver() != null) {
       getFlowReceiver().close();
+    }
     super.close();
   }
 
-  @SuppressWarnings("deprecation")
   @Override
   public void prepare() throws CoreException {
     Args.notBlank(getQueue(), "queue");
@@ -94,4 +98,5 @@ public class SolaceJcsmpQueueConsumer extends SolaceJcsmpAbstractConsumer {
   private String queueName() {
     return getQueue();
   }
+
 }
