@@ -6,7 +6,7 @@ import org.slf4j.LoggerFactory;
 import com.solacesystems.jcsmp.BytesXMLMessage;
 import com.solacesystems.jcsmp.JCSMPException;
 import com.solacesystems.jcsmp.JCSMPSession;
-import com.solacesystems.jcsmp.JCSMPStreamingPublishEventHandler;
+import com.solacesystems.jcsmp.JCSMPStreamingPublishCorrelatingEventHandler;
 import com.solacesystems.jcsmp.XMLMessageProducer;
 import com.solacesystems.jcsmp.transaction.TransactedSession;
 import com.solacesystems.jcsmp.transaction.TransactionStatus;
@@ -15,63 +15,65 @@ import lombok.Getter;
 import lombok.Setter;
 
 public class SolaceJcsmpSessionHelper {
-  
+
   protected transient Logger log = LoggerFactory.getLogger(this.getClass().getName());
 
   @Getter
   @Setter
   private boolean transacted;
-  
+
   @Getter
   @Setter
   private TransactedSession transactedSession;
-  
+
   @Getter
   @Setter
   private JCSMPSession session;
-  
+
   @Getter
   @Setter
   private SolaceJcsmpConnection connection;
-  
+
   public SolaceJcsmpSessionHelper() {
   }
-  
+
   private void createStandardSession() throws Exception {
     setSession(getConnection().createSession());
   }
-  
+
   private boolean isStandardSessionActive() {
-    return getSession()  != null && !getSession().isClosed();
+    return getSession() != null && !getSession().isClosed();
   }
-  
+
   private boolean isTransactedSessionActive() {
-    return getTransactedSession()  != null && !(getTransactedSession().getStatus() == TransactionStatus.CLOSED);
+    return getTransactedSession() != null && !(getTransactedSession().getStatus() == TransactionStatus.CLOSED);
   }
-  
+
   private void createTransactedSession() throws Exception {
-    if(!isStandardSessionActive())
+    if (!isStandardSessionActive()) {
       createStandardSession();
+    }
     setTransactedSession(getSession().createTransactedSession());
   }
-  
+
   boolean isSessionActive() {
     return getTransacted() ? isTransactedSessionActive() : isStandardSessionActive();
   }
-  
-  XMLMessageProducer createMessageProducer(JCSMPStreamingPublishEventHandler eventHandler) throws JCSMPException {
+
+  XMLMessageProducer createMessageProducer(JCSMPStreamingPublishCorrelatingEventHandler eventHandler) throws JCSMPException {
     return getSession().getMessageProducer(eventHandler);
   }
-  
+
   void createSession() throws Exception {
-    if(getTransacted())
+    if (getTransacted()) {
       createTransactedSession();
-    else
+    } else {
       createStandardSession();
+    }
   }
-  
+
   void commit(BytesXMLMessage message) {
-    if(getTransacted()) {
+    if (getTransacted()) {
       try {
         getTransactedSession().commit();
         log.trace("Committing transaction.");
@@ -79,30 +81,30 @@ public class SolaceJcsmpSessionHelper {
         log.error("Commit failed, rolling back.", e);
         rollback();
       }
-    }
-    else
+    } else {
       message.ackMessage();
+    }
   }
-  
+
   void rollback() {
-    if(getTransacted()) {
+    if (getTransacted()) {
       try {
         getTransactedSession().rollback();
       } catch (JCSMPException e) {
         log.error("Rollback failed.", e);
       }
-      
+
     }
   }
-  
+
   void close() {
-    if(getTransacted()) {
-      if(isTransactedSessionActive())
+    if (getTransacted()) {
+      if (isTransactedSessionActive()) {
         getTransactedSession().close();
-    } else {
-      if(isStandardSessionActive())
-        getSession().closeSession();
+      }
+    } else if (isStandardSessionActive()) {
+      getSession().closeSession();
     }
-      
   }
+
 }
